+++ graphics/java/android/graphics/Xfermode.java
@@ -38,7 +38,7 @@ public class Xfermode {
         }
     }
 
-    private static native void finalizer(int native_instance);
+    private static native void finalizer(long native_instance);
 
-    int native_instance;
+    long native_instance;
 }
